LPD-82344 Refactor and move generated tests into existing files
